moving scripts to app folder, but still has errors in models
